backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
PyMuPDF>=1.22.0
ollama>=0.1.0
python-docx>=0.8.11
python-dotenv>=1.0.0